requirements.txt
pydantic
pygrocy
pyserial
